Frontend-Backend/text-to-sql-app/requirements.txt
# Core dependencies
flask>=2.0.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
requests>=2.25.0

# BigQuery integration
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0

# Basic utilities
numpy>=1.20.0
pandas>=1.3.0

# Testing
pytest>=7.0.0